+++ database/migrations/2021_01_05_081915_add_cascade_delete_to_comments_table.php
@@ -21,10 +21,13 @@ class AddCascadeDeleteToCommentsTable extends Migration
         });
 
         Schema::table('comments', function (Blueprint $table) {
-            $table->foreignId('blog_post_id')
-                ->default(0)
-                ->constrained()
-                ->onDelete('cascade');
+            if (env('DB_CONNECTION') !== 'sqlite_testing') {
+                $table->unsignedInteger('blog_post_id');
+                $table->foreign('blog_post_id')
+                    ->references('id')
+                    ->on('blog_posts')->default(0)
+                    ->onDelete('cascade');
+            }
         });
     }
 
@@ -35,14 +38,19 @@ class AddCascadeDeleteToCommentsTable extends Migration
      */
     public function down()
     {
+
         Schema::table('comments', function (Blueprint $table) {
             $table->dropForeign(['blog_post_id']);
-            $table->dropColumn('blog_post_id');
+            $table->dropIfExists('blog_post_id');
         });
+
         Schema::table('comments', function (Blueprint $table) {
-            $table->foreignId('blog_post_id')
-                ->default(0)
-                ->constrained();
+            if (Schema::hasColumn('comments', 'blog_post_id')) {
+                $table->unsignedInteger('blog_post_id');
+                $table->foreign('blog_post_id')
+                    ->references('id')
+                    ->on('blog_posts')->default(0);
+            }
         });
     }
 }
